Filter ambiguous alignments
I added a filter to check whether the insert, when realigned
for non-exact ITD search, is realigned in one piece or split.
These split alignments seem to be the only ambiguous alignmments
detected for MOLM14-1-minBQS_35. For the future, statistics on
this are output so I will know whether this is something I need
to take care of for other samples.