Annotate domains & coordinates,  other stuff?
I added annotation of genomic, transcriptomic and proteomic
coordinates as well as known FLT3 domains and exons.

I failed to commit before doing so, so there are old changes
present as well... Maybe sth about fixing trailing? Trailing muts
should only be considered special when they occur in the middle of
the amplicon. At the very beginning or end they should be treated
like regular ITDs/inserts. (Not sure that was it / was all though)