Require tandem2 to reside within read
For non-trailing ITDs this is implicitely required by requiring
adjacency: Neighboring sequences of non-trailing ITD are always
covered by the read, otherwise the insert would not be non-trailing.

For trailing ITDs with trailing_end == 5:
	require tandem2_start > start
For trailing ITDs with trailing_end == 3:
	require tandem2_start < start

For true ITDs, tandem2 will actually always reside within the read:
Because ITDs are internal TDs, at least one tandem is part of the
reference sequence and mapped to it. Since the insert is the tandem
that is not aligned / part of the reference, it must be the second
tandem. It is therefore safe to require that this reside within
the supporting read in all cases.