Right-shift insertions
Previously, when aligning read to ref, I always selected the first
of the list of optimal alignments. There, the insert is maximally
left-shifted and the ITD insert typically preceded the WT tandem.

Laura and Frank Rücker want this the other way around, so I now
pick the last alignment of the list.