Fix insertion site annotation
Previously, I annotated the bp succeeding the ITD, which may or
may not be part  of the same AS. I need to annotate the succeeding
AS instead though.

By annotating the AS of (last-ITD-bp + 3), I should always do so.
I removed annotation of insertion site coordinate on genomic and
transcriptomic level as this does not make sense with this change
and also is not necessary.