Fix merging -> Add multiple rounds of merging
Merging was previously aborted as soon as a specific
InsertCollection was merged to any other. However,
sometimes this InsertCollection was exactly the one
required to merge two other Collections:

Collection A and Collection B do not merge with each other
but C merges with both
-->  previouly C would merge with A or B, depending on which
was tested first but AC and B would remain separate.

I now test for each Collection whether there is any other /
additional collection to merge. If there is, I perform
a whole additional round of merging on the same condition.

This solves the problem of incomplete merging and makes
success of merging independent of the order of Collections
during merging and thus chance.