Reimplement coverage calculation in separate loop
Coverage was updated for each read within the loop
wherein also reads where checked for whether they
contained any insertions.

It's cleaner to do coverage calculation across the
reference in a separate loop / section of code.

This was also necessary to be able to adjust coverage
calculation later to do it on a read-pair instead of
single-read basis.