Make reverse_complement() non-destructive
Although class methods return object, input "self" was changed
in-place as well. This is a problem with 454 processing, where
both forward and reverse-complemented reads are aligned to the
reference and the better fit is selected
--> when the original read was chosen, it was nontheless
already reverse-complemented!

(Besides this return-object-but-additionally-change-it-in-place
being problematic here, it is also confusing and counter-intuitive.
I should probably change this at some point!?)