Rewrite insertion_site and domain annotation
To enable testing of insertion_site and domain
annotation, I needed methods to explicitely fetch
these for Insertion / ITD instances (instead of pandas
DataFrame about to be written to file) independent of /
without writing anything to file.

This change also fixes a previous bug that I wasn't aware of:
Domains were annotated to not-prepped-for-save insertions,
which means that start/end coordinates were still 0-based
instead of 1-based as the annotation is and, most importantly,
for ITDs they were still insert coordinates instead of (WT)
tandem2 coords!