Fix unaligned-primer filter
Filter alignments with gaps in primers.

Previously I was filtering alignments with deletions in the primer
sequence but not insertions. This makes no sense so to be consistent
I am filtering indels in general.

Also, in-code comments and output printed mentioned filtering reads
with more than 3 gaps but in fact I was filtering all deletions.