Save / Load config file for multiprocessing on Windows
Commandline arguments were previously simply stored in a global
dict whose values were filled at the beginning of __main__.

On Windows, child processes are spawned without receiving
a copy of the parent process'. Because the empty config
was declared at the beginning of the module and child
processes do not (and should  not) loop over the __main__
themselves, they had access only to the empty config dict,
causing the program to fail.

I am now saving the config to file within __main__
and loading it from that file for parallel processes.

Retrieval of the correct config file requires that I
enter the respective output directory. To not break
any file and directory paths provided by the user,
I make sure these are all absolute.