Refine argparse to make R2 fastq optional
__main__ code was already testing for whether the R2 file
was provided before processing it, but argparse still
required it to be provided. I made it an optional
positional argument, had to change the order of args
for this to work though.

Also added other minor changes: Changed order of args in
the code, renamed minBQS to min_bqs for consistency and
added "choices=['Illumina', '454']" to TECH arg to
have argparse enforce this (instead of simply telling
user that these are the only two sensible choices so far).