Add MAX_TRAILING_BP to recognize "nearly-trailing" inserts
Some trailing insertions were not recognized as such when
by chance the last insert base sequenced matched the first
WT reference base subsequent to it. These inserts then
failed to be recognized as ITDs because they were either
not in-frame or not adjacent to their WT tandem.

For example below, the final "G" should not be mapped. Because it is mapped, the insertion
is no longer truly at the very 3' end of the read and therefore not actually "trailing".
To compensate, and recognize this insertion as the ITD that it is, config["MAX_TRAILING_BP"]
can be set > 0 to recognize insertions with at most X trailing bp mapped as "trailing"
anyway.

    1 GAAATTTAGGTATGAAAGCCAGCTACAGATGGTACAGGTGACCGGCTCCT     50
      |.||||||||||||||||||||||||||||||||||||||||||||||||
    1 GCAATTTAGGTATGAAAGCCAGCTACAGATGGTACAGGTGACCGGCTCCT     50

   51 CAGATAATGAGTACTTCTACGTTGATTTCAGAGAATATGAATATGATCTC    100
      ||||||||||||||||||||||||||||||||||||||||||||||||||
   51 CAGATAATGAGTACTTCTACGTTGATTTCAGAGAATATGAATATGATCTC    100

  101 AAATGGGAGTTTCCAAGAGAAAATTTAGAGTTTGGTAAGAATGGAATGTG    150
      ||||||||||||||||||||||||||||||||||||||||||||||||||
  101 AAATGGGAGTTTCCAAGAGAAAATTTAGAGTTTGGTAAGAATGGAATGTG    150

  151 CCAAATGTTTCTGCAGCATTTCTTTTCCATTGGAAAATCTTTAAAATGCA    200
      ||||||||||||||||||||||||||||||||||||||||||||||||||
  151 CCAAATGTTTCTGCAGCATTTCTTTTCCATTGGAAAATCTTTAAAATGCA    200

  201 CGTACTCACCATTTGTCTTTGCAGGGAAGCCACAGGTGACCGGCTCCTCA    250
      |||||||||||||||||||||||||||||
  201 CGTACTCACCATTTGTCTTTGCAGGGAAG---------------------    229

  251 G-------------------------------------------------    251
      |
  230 GTACTAGGATCAGGTGCTTTTGGAAAAGTGATGAACGCAACAGCTTATGG    279

  252 --------------------------------------------------    251

  280 AATTAGCAAAACAGGAGTCTCAATCCAGGTTGCCGTCAAAATGCTGAAAG    329

I also added the respective test to the test_getitd.py file.