Set max-trailing-bp to 0  -->  disable feature
With this set, additional ITDs (small trailing, probably FPs) are
found. That wasn't the point when I set out to fix ID 264 dx VAF,
so I am disabling this feature for now. We do not want
additional false positives... one can use the "insertion" result
files if need be / to check whether something went wrong with
the respective ITD