Fix getitd_from_config.py: properly parse bools
All arguments were read in as strings and bools were not
recognized as such. This caused `"False"` to be later
evaluated to `True`, messing up getITD configuration.
Mind that this had previously been fixed for the main
script parsing arguments from the command line but
not the getitd_from_config.py helper script (see
commit 329c5485a1cedb3efd9e169dd300ea91556219d2).

Now, as in the main script, respective strings
are explicitely parsed as booleans.