Filter ITDs with insertion_site within WT tandem
I found a case where a short trailing insert could be mapped
to two locations within the reference, one of which was valid
and one of which overlapped the actual insert's insertion site,
which is not valid. Alignments are now filtered to remove any
such invalid ones before one is chosen to call an ITD from.