Implement inter-bp coverage
Previously, coverage of each reference bp was defined as the sum
of sequencing reads aligned to the respective position. Inserts'
coordinates within the reference were defined as the preceding WT
base of the reference, and from this base was then the reference
coverage extracted. If now a single read was aligned which contained
a 5' insert, then the VAF of this insert was calculated as
1 (= # of insert-supporting reads) / 0 (= 3 of reads covering the
preceding WT base). In other cases, VAF would come out as > 100%
or more or less incorrect, without any error being thrown.

To fix this, I switched reference coverage calculation to that of
inter-bp: An inter-bp coordinate has its coverage increased when
a given read spans across it. This also helps to distinguish reads
whose alignment ends on an insertion site from those that span it.
The former should not be included in coverage calculation for the
respective insert because it is unknown whether it would have
supported it or not.