Extend annotation files
One sample failed when 6bp of its 7bp 5' trailing insert mapped
to the very 3' end of the read, coinciding with the 3' end of the
reference: The 7th basepair led to bp 331 being annotated as the
insertion's end, for which subsequent annotation failed with
index out of range.

Not sure this is the best fix. Are there cases that would require
further extension (at either end)? Should the ITD have been called?
Technically the WT tandem wasn't fully sequenced, if we assume this
7th bp was part of it as well.