Rm (wrong) check that WT tandem is fully sequenced
The way it was written, especially 5' trailing were not necessarily
handled correctly. Standard ITDs should not need this requirement,
but it can remove some of the weird artefacts we see at times.

Might have to be added back in - to evaluate this better, reads
with incomplete WT tandems are written to a separate file in the
getITD output folder, named "incomplete-wt-tandem.log".