Add `-min_insert_seq_length` param
Add an optional parameter to control the minimum length required
for an insert to be considered and called by getITD. The default 6
seems to work well for high-quality data, but low-quality samples
are prone to short, spurious trailing ITDs with 6-7 bp inserts.
This parameter should allow better filtering of such cases.